+++ apps/quick-zoom/src/app/app.tsx
@@ -16,6 +16,7 @@ import {
   ZoomLinkProps,
 } from './zoom-link/zoom-link'
 import { ImportExport } from './import-export/import-export'
+import { fromJson } from './from-json/from-json'
 
 const useStyles = makeStyles((theme) =>
   createStyles({
@@ -65,32 +66,6 @@ const reducer = (
     }
   }
   return state
-}
-
-export const fromJson = (json: string): ZoomLinkProps[] => {
-  const items = JSON.parse(json)
-
-  if (!Array.isArray(items)) {
-    return []
-  }
-
-  return items.map(
-    ({
-      name,
-      url: link,
-      confno,
-      pwd,
-    }: Omit<ZoomLinkProps, 'url'> & { url: string }) => {
-      const url = new URL(link)
-
-      if (confno !== undefined) {
-        return { name, url, confno, pwd }
-      }
-
-      const [confnoFromUrl, pwdFromUrl] = fromUrl(url)
-      return { name, url, confno: confnoFromUrl, pwd: pwdFromUrl }
-    },
-  )
 }
 
 const loadFromStorage = (): ZoomLinkProps[] => {
